Simplified implementation of TableIterator.
Abolished InvalidTableIterator exception. It is now
undefined behaviour to dereference a null TableIterator.
This is consistent with the standard library iterators.
However, incrementing a null TableIterator is now fine - it simply
resets the underlying SQLStatement and cycles again through the
result set. Incrementing a default-constructed TableIterator
(with which no result set is associated) is simply a no-op.

Added copy constructor and postfix increment to TableIterator.
Wrote tests for these. Tests pass.

Made TableIterator inherit from std::iterator<...>, with template
arguments to std::iterator appropriate for an input iterator.
TableIterator should now work with the standard library algorithms
that are appropriate to input iterators (and cursory experimentation
suggests strongly that it does indeed work with these).



git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/sqloxx@2157 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98