Provided the ability for IdentityMap and sqloxx::Handle to be used to manage persistent objects with a dynamic type that is different from their static type, such that the static type provides the IdentityMap and primary key, while the dynamic type is derived from the static type but shares its primary key.
The modifications to allow this include the creation of a PersistenceTraits
class, which provides the "primary type" (PersistenceTraits<T>::PrimaryT) for
a given type T (which defaults to T itself). The primary type is the type
such that its primary key column provides the primary key for T. The
IdentityMap for T will be the IdentityMap for the primary type of T.
T should either be, or be derived from, T.

These changes should facilitate the handling of persistent type hierarchies in
client code, which previously was quite awkward.

It is likely further refinements will be required to bed down this
significant expansion of functionality for SQLoxx.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/sqloxx/trunk@2300 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98